backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
celery==5.6.3
channels==4.3.2
channels-redis==4.3.0
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
factory-boy==3.3.3
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Markdown==3.10.2
psycopg2==2.9.11
PyYAML==6.0.3
pytest==9.0.3
pytest-django==4.12.0
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2026.1
uritemplate==4.2.0
